arreglada dashboard para cuando el populate esta vacio

+++ Item 4/Acme BnB/src/main/java/repositories/InvoiceRepository.java
@@ -12,18 +12,18 @@ public interface InvoiceRepository extends JpaRepository<Invoice, Integer> {
 
 	//Returns the minimum number of invoices per actor
 	@Query("select min(t.invoices.size) from Tenant t")
-	public double getMinimumInvoicesPerTenant();
+	public Integer getMinimumInvoicesPerTenant();
 
 	//Returns the average number of invoices per actor
 	@Query("select avg(t.invoices.size) from Tenant t")
-	public double getAverageInvoicesPerTenant();
+	public Double getAverageInvoicesPerTenant();
 
 	//Returns the maximum number of invoices per actor
 	@Query("select max(t.invoices.size) from Tenant t")
-	public double getMaximumInvoicesPerTenant();
+	public Integer getMaximumInvoicesPerTenant();
 
 	//Returns the total amount of money due in the invoices issued by the system
 	@Query("select sum(b.totalAmount) from Book b where b.invoice is not null")
-	public double getTotalDueMoneyOfInvoices();
+	public Double getTotalDueMoneyOfInvoices();
 
 }
